Fix: sessiond: invalid session configuration on EXCLUDE_ALL policy
Saving a session with a process attribute tracker that uses the
`EXCLUDE_ALL` policy results in an invalid session configuration.

Currently, a tracker of the following form is produced:
<process_attr_values>
  <vpid/>
</process_attr_values>

This is invalid as per the XSD as 'vpid' is not a list; it is an
individual tracked attribute.

The appropriate '<process_attr_values/>' empty node is now produced.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Ia31e276fe7246a89e72d6808c9ed960fb04f1b3a